pyproject.toml
[tool.poetry]
name = "last-sprint"
version = "0.1.0"
description = ""
authors = ["zormnus <zotov.max17@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
channels = "^4.0.0"
daphne = "^4.1.0"
channels-redis = "^4.2.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
